+++ src_3_0_0/modules/alerting/store/actionsAlerting.js
@@ -120,6 +120,7 @@ export default {
 
         let category,
             isUnique = false,
+            onceInSession = false,
             isNotRestricted = false,
             isInTime = false,
             displayAlert = false;
@@ -166,6 +167,7 @@ export default {
         }
         alertProtoClone.hash = objectHash(alertProtoClone.hash);
         isUnique = findSingleAlertByHash(state.alerts, alertProtoClone.hash) === false;
+        onceInSession = !isUnique ? alertProtoClone.onceInSession : false;
         if (!isUnique) {
             console.warn("Alert ignored (duplicate): " + alertProtoClone.hash);
         }
@@ -191,11 +193,14 @@ export default {
         }
         // even if current alert got seeded out, there still might be another one in the pipe
         if (state.alerts.length > 0) {
-            if (findSingleAlertByHash(state.alerts, alertProtoClone.hash) !== false && isInTime && isNotRestricted) {
+            if (findSingleAlertByHash(state.alerts, alertProtoClone.hash) !== false && isInTime && isNotRestricted && !onceInSession) {
                 // this is necessary because this action returned false even if the alert was displayed
                 displayAlert = true;
             }
-            commit("setReadyToShow", true);
+
+            if (displayAlert) {
+                commit("setReadyToShow", true);
+            }
         }
 
         return displayAlert;