+++ src_3_0_0/modules/alerting/store/gettersAlerting.js
@@ -30,6 +30,7 @@ export default {
 
             return sortingArr.indexOf(b.category) - sortingArr.indexOf(a.category);
         });
+
         return results;
     }
 };