+++ src_3_0_0/modules/alerting/tests/unit/store/actionsAlerting.spec.js
@@ -5,7 +5,6 @@ import {expect} from "chai";
 describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
     let commit;
 
-
     beforeEach(() => {
         commit = sinon.spy();
     });
@@ -14,7 +13,7 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         sinon.restore();
     });
 
-    it("cleanup", async function () {
+    it("cleanup", () => {
         const state = {
             alerts: [
                 {
@@ -31,7 +30,8 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.getCall(1).args).to.eql(["removeFromAlerts", {hash: "123", initial: true, mustBeConfirmed: false}]);
         expect(commit.getCall(2).args).to.eql(["setReadyToShow", false]);
     });
-    it("setAlertAsRead", async function () {
+
+    it("setAlertAsRead", () => {
         const state = {
             alerts: [
                 {
@@ -45,7 +45,8 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.calledOnce).to.be.true;
         expect(commit.firstCall.args).to.eql(["setAlertAsRead", {hash: "found", mustBeConfirmed: true}]);
     });
-    it("setAlertAsUnread", async function () {
+
+    it("setAlertAsUnread", () => {
         const state = {
             alerts: [
                 {
@@ -59,7 +60,8 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.calledOnce).to.be.true;
         expect(commit.firstCall.args).to.eql(["setAlertAsUnread", {hash: "found", mustBeConfirmed: false}]);
     });
-    it("addSingleAlert adds a valid alert", async function () {
+
+    it("addSingleAlert adds a valid alert", () => {
         const state = {
                 alerts: [],
                 displayedAlerts: [],
@@ -72,7 +74,8 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.firstCall.args[0]).to.eql("Modules/News/addNews");
         expect(commit.firstCall.args[1].content).to.eql("123");
     });
-    it("addSingleAlert doesnt show alert with not valid time restriction", async function () {
+
+    it("addSingleAlert doesnt show alert with not valid time restriction", () => {
         const state = {
                 alerts: [],
                 displayedAlerts: [],
@@ -85,7 +88,8 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.firstCall.args[0]).to.eql("Modules/News/addNews");
         expect(commit.firstCall.args[1].content).to.eql("123");
     });
-    it("addSingleAlert shows alert with valid time restriction", async function () {
+
+    it("addSingleAlert shows alert with valid time restriction", () => {
         const state = {
                 alerts: [],
                 displayedAlerts: [],
@@ -97,9 +101,9 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.calledTwice).to.be.true;
         expect(commit.firstCall.args[0]).to.eql("Modules/News/addNews");
         expect(commit.firstCall.args[1].content).to.eql("123");
-
     });
-    it("addSingleAlert doesnt show alert with already existing hash", async function () {
+
+    it("addSingleAlert doesnt show alert with already existing hash", () => {
         sinon.stub(console, "warn");
         const state = {
             alerts: [{hash: "2db391b6db7866773874b3a5e60123040adb656a"}],
@@ -116,5 +120,31 @@ describe("src_3_0_0/modules/alerting/store/actionsAlerting.js", () => {
         expect(commit.firstCall.args[1].content).to.eql("123");
         expect(commit.secondCall.args[0]).to.eql("setReadyToShow");
         expect(commit.secondCall.args[1]).to.be.true;
+    });
+
+    it("should only execute addNews if onceInSession === true and new alert already exists", () => {
+        const state = {
+            alerts: [{
+                hash: "2db391b6db7866773874b3a5e60123040adb656a",
+                content: "123",
+                displayFrom: "2022-08-24 05:00",
+                displayUntil: "2088-09-28 23:59",
+                onceInSession: true
+            }],
+            displayedAlerts: [],
+            availableCategories: ["news", "success", "warning", "error", "info"]
+        };
+
+        actions.addSingleAlert({state, commit}, {
+            hash: "2db391b6db7866773874b3a5e60123040adb656a",
+            content: "123",
+            displayFrom: "2022-08-24 05:00",
+            displayUntil: "2088-09-28 23:59",
+            onceInSession: true
+        });
+
+        expect(commit.calledOnce).to.be.true;
+        expect(commit.firstCall.args[0]).to.eql("Modules/News/addNews");
+        expect(commit.firstCall.args[1].content).to.eql("123");
     });
 });