+++ src_3_0_0/modules/layerTree/tests/unit/components/LayerCheckBox.spec.js
@@ -41,6 +41,12 @@ describe("src_3_0_0/modules/layerTree/components/LayerCheckBox.vue", () => {
         store = createStore({
             namespaces: true,
             modules: {
+                Alerting: {
+                    namespaced: true,
+                    actions: {
+                        addSingleAlert: sinon.stub()
+                    }
+                },
                 Modules: {
                     namespaced: true,
                     modules: {
@@ -61,6 +67,14 @@ describe("src_3_0_0/modules/layerTree/components/LayerCheckBox.vue", () => {
             },
             actions: {
                 replaceByIdInLayerConfig: replaceByIdInLayerConfigSpy
+            },
+            getters: {
+                layerConfigById: () => () => {
+                    return {
+                        layerAttribution: "This is the layer attribution!",
+                        name: "The layer"
+                    };
+                }
             }
         });
     });
@@ -295,5 +309,26 @@ describe("src_3_0_0/modules/layerTree/components/LayerCheckBox.vue", () => {
         expect(removeSelectedLayerSpy.firstCall.args[1]).to.be.deep.equals(spyArg);
     });
 
+    describe("showLayerAttributions", () => {
+        it("should throw an alert, if layer has a layerAttribution", () => {
+            wrapper = shallowMount(LayerCheckBox, {
+                global: {
+                    plugins: [store]
+                },
+                propsData
+            });
 
+            const addSingleAlertSpy = sinon.spy(wrapper.vm, "addSingleAlert");
+
+            wrapper.vm.showLayerAttributions(true);
+
+            expect(addSingleAlertSpy.calledOnce).to.be.true;
+            expect(addSingleAlertSpy.firstCall.args[0]).to.deep.equals({
+                content: "This is the layer attribution!",
+                category: "info",
+                title: "The layer",
+                onceInSession: true
+            });
+        });
+    });
 });