+++ CHANGELOG.md
@@ -132,6 +132,11 @@
 - WFS: Migrated writeTransaction and sendTransaction to masterportalAPI (2.22.0).
 - GFI: gfiFeatures are no longer set to null on every click, but only when there are no features or the gfi is closed
 
+### Deprecated
+
+### Removed
+- The loader start screen has been removed.
+
 ### Fixed
 - Print: The portal is no longer reloaded after pressing the Enter key in the title field. Instead, the print is sent.
 - WFST-Tool:
@@ -229,6 +234,7 @@
 - You can configure the posiibility to add attachments to the `contact` form.
 - Added documentation file doc\jsdoc.md for jsdoc in vue-components and vuex-files.
 - Script to migrate masterportal configuration files to version 3.0.0
+- The layer attribution is now displayed in an alert when a layer is activated for the first time
 
 ### Added
 - GFI: nested object values can now be addressed from "gfiAttributes" by dot notation. See [services.json](https://bitbucket.org/geowerkstatt-hamburg/masterportal/src/dev/doc/services.json.md#markdown-header-gfi_attributes) .